+++ indexer/categories_index.cpp
@@ -52,7 +52,7 @@ namespace indexer
 {
 void CategoriesIndex::AddCategoryByTypeAndLang(uint32_t type, int8_t lang)
 {
-  ASSERT(lang >= 1 && lang <= static_cast<int8_t>(CategoriesHolder::kLocaleMapping.size()),
+  ASSERT(lang >= 1 && static_cast<size_t>(lang) <= CategoriesHolder::kLocaleMapping.size(),
          ("Invalid lang code:", lang));
   m_catHolder->ForEachNameByType(type, [&](TCategory::Name const & name)
                                  {
@@ -69,7 +69,7 @@ void CategoriesIndex::AddCategoryByTypeAllLangs(uint32_t type)
 
 void CategoriesIndex::AddAllCategoriesInLang(int8_t lang)
 {
-  ASSERT(lang >= 1 && lang <= static_cast<int8_t>(CategoriesHolder::kLocaleMapping.size()),
+  ASSERT(lang >= 1 && static_cast<size_t>(lang) <= CategoriesHolder::kLocaleMapping.size(),
          ("Invalid lang code:", lang));
   m_catHolder->ForEachTypeAndCategory([&](uint32_t type, TCategory const & cat)
                                       {
